+++ noro/noropriv.h
@@ -40,7 +40,7 @@
 typedef struct netpoller netpoller;
 netpoller* netpoller_new();
 void netpoller_loop();
-void netpoller_yieldfd(int fd, int ytype, void* gr);
+void netpoller_yieldfd(long fd, int ytype, void* gr);
 void netpoller_use_threads();
 
 // for goroutine
@@ -72,8 +72,8 @@ struct goroutine {
 };
 
 // processor callbacks, impl in noro.c
-extern int noro_processor_yield(int fd, int ytype);
-extern int noro_processor_yield_multi(int ytype, int nfds, int fds[], int ytypes[]);
+extern int noro_processor_yield(long fd, int ytype);
+extern int noro_processor_yield_multi(int ytype, int nfds, long fds[], int ytypes[]);
 extern bool noro_in_processor();
 extern void noro_processor_resume_some(void* gr_, int ytype);
 extern goroutine* noro_goroutine_getcur();
